+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "ic-repl"
-version = "0.3.3"
+version = "0.3.4"
 authors = ["DFINITY Team"]
 edition = "2018"
